Cleanup, fix some warnings and move some more defines to stdafx.h for simpler tweaking
git-svn-id: https://svn.code.sf.net/p/vpinball/code/trunk@344 14438d79-0ffd-4333-8ae7-ea2bdd901387

+++ pinsound.cpp
@@ -86,7 +86,7 @@ void PinDirectSound::InitDirectSound(HWND hwnd)
 	if (!FAILED (DirectSoundEnumerate (DSEnumCallBack, &DSads)))
 	{
 		hr = GetRegInt("Player", "SoundDevice", &DSidx);
-		if ((hr != S_OK) || (DSidx >= DSads.size()))
+		if ((hr != S_OK) || ((unsigned int)DSidx >= DSads.size()))
 		{
 			DSidx = 0; // The default primary sound device
 		}
